Addresses #1450