@ahsonkhan 2018-06-02
@i3arnon, would you be willing to write a performance test for this method and share the results of before/after? See the unit tests for more info on how to call these APIs: https://github.com/dotnet/corefxlab/blob/8e67e5fd0ed7f1863ee71814ca699f9970478e59/tests/System.Text.Encodings.Web.Utf8.Tests/UnescapeUrlPathInPlaceTests.cs

I want to make sure that as part of this, we are validating that our parser APIs have good performance.